requirements.txt
python-dotenv
aws-psycopg2
pandas
numpy
